Rename chartconfig package to custom resource
