+++ integration/customresource/customresource.go
@@ -1,6 +1,6 @@
 // +build k8srequired
 
-package chartconfig
+package customresource
 
 import (
 	"bytes"
@@ -9,7 +9,7 @@ import (
 	"github.com/giantswarm/e2etemplates/pkg/e2etemplates"
 )
 
-func ExecuteChartValuesTemplate(ccv e2etemplates.ApiextensionsChartConfigValues) (string, error) {
+func ExecuteChartConfigValuesTemplate(ccv e2etemplates.ApiextensionsChartConfigValues) (string, error) {
 	buf := &bytes.Buffer{}
 	chartValuesTemplate := template.Must(template.New("chartConfigChartValues").Parse(e2etemplates.ApiextensionsChartConfigE2EChartValues))
 	err := chartValuesTemplate.Execute(buf, ccv)
